PR 10a step 1 — lift boot_registry into new lex-engine crate
The boot helper that turns CLI/host inputs into a populated Registry
+ trust gate moves out of lex-cli into a new lex-engine crate. lex-cli
keeps a thin shim (CliPromptHandler + parameter forwarding) so existing
call sites at main.rs:597 and the labels_subcommand tests don't budge.
The lex-lsp wiring landing next can call the same lex-engine helper
without depending on the CLI binary crate.

The lifted helper takes a Box<dyn TrustPromptHandler> on the setup
struct so each host installs the prompt that fits its UX:
- CLI: CliPromptHandler (denies with --enable-handlers rationale).
- LSP (10a): denies with "use lexd labels in CLI mode first" rationale.
- LSP (10b): forwards lex/trustRequest to the editor.
- Embedders (PR 11): inject whatever fits.

The new crate sits above lex-core in the dep graph (it needs
builtins::register_into for the lex.* native handlers); it can't live
in lex-extension-host because lex-core already depends on that. The
public Engine::builder facade planned in PR 11 will live here too.
